@@ -10,17 +10,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.11",
+    "@angular/animations": "^14.0.0",
     "@angular/cdk": "^13.3.9",
-    "@angular/common": "~13.3.11",
-    "@angular/compiler": "~13.3.11",
-    "@angular/core": "~13.3.11",
+    "@angular/common": "^14.0.0",
+    "@angular/compiler": "^14.0.0",
+    "@angular/core": "^14.0.0",
     "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "~13.3.11",
+    "@angular/forms": "^14.0.0",
     "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "~13.3.11",
-    "@angular/platform-browser-dynamic": "~13.3.11",
-    "@angular/router": "~13.3.11",
+    "@angular/platform-browser": "^14.0.0",
+    "@angular/platform-browser-dynamic": "^14.0.0",
+    "@angular/router": "^14.0.0",
     "@ngneat/until-destroy": "^9.0.0",
     "@ngrx/effects": "^13.2.0",
     "@ngrx/store": "^13.2.0",
@@ -34,8 +34,8 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^14.0.0",
-    "@angular/cli": "~13.3.7",
-    "@angular/compiler-cli": "~13.3.11",
+    "@angular/cli": "^14.0.0",
+    "@angular/compiler-cli": "^14.0.0",
     "@ngneat/until-destroy-migration": "^9.0.0",
     "@testing-library/angular": "^11.0.2",
     "@types/assert": "^1.5.6",
